backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
cloudinary==1.34.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
Werkzeug==2.3.7